Backend(FastAPI)/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pymongo
streamlit
deep-translator
spacy
